image/requirements.txt
langchain == 0.3.26
fastapi == 0.116.1
mangum == 0.19.0
uvicorn == 0.35.0
requests == 2.32.4
python-dotenv==1.1.1
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
python-multipart==0.0.20